+++ app/src/main/java/org/example/domain/login/LoginService.java
@@ -3,8 +3,8 @@ package org.example.domain.login;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpSession;
 import lombok.extern.slf4j.Slf4j;
-import org.example.domain.login.dto.LoginForm;
-import org.example.domain.login.dto.LoginResponse;
+import org.example.domain.login.dto.request.LoginForm;
+import org.example.domain.login.dto.response.LoginResponse;
 import org.example.domain.member.MemberRepository;
 import org.example.domain.member.entity.Member;
 import org.example.domain.session.SessionConst;
